metastore: working examples for testing with mongodb and mongoose
+ refactoring
