+++ services/metastore/.env
@@ -1,4 +1,4 @@
-metastore_port=3003
-mongodb_port=27017
+METASTORE_PORT=3003
+MONGODB_PORT=27017
 
 COMPOSE_PROJECT_NAME=potpourri-metastore
