+++ services/metastore/index.js
@@ -4,6 +4,8 @@
  * This script starts the server for production.
  */
 
+require('dotenv').config()
+
 const server = require('./server')
 
-server.initDB()
+server.initDB('mongodb://localhost:' + process.env.MONGODB_PORT + '/pot-pourri')
