+++ services/metastore/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "mongoose": "^5.9.24",
     "multer": "^1.4.2"
@@ -27,12 +28,14 @@
     "chai": "^4.2.0",
     "chai-fs": "^2.0.0",
     "chai-http": "^4.3.0",
+    "chai-like": "^1.1.1",
     "eslint": "^7.3.1",
     "eslint-plugin-chai-friendly": "^0.6.0",
     "eslint-plugin-mocha": "^7.0.1",
     "eslint-plugin-standard": "^4.0.1",
     "mocha": "^8.0.1",
     "mock-fs": "^4.12.0",
+    "mongodb": "^3.5.9",
     "mongodb-memory-server": "^6.6.1",
     "nodemon": "^2.0.4",
     "nyc": "^15.1.0",
